Allow for custom JS variable name (for global)
The current isset($this->chart->renderTo) is returning false even when set, so the chart is always being set on the variable named chart (in JS).  

In our application we need to be able to set the chart on a global variable so we can access it from other scripts in order to call Highchart's exportChart() method from JS listeners, and in order to write custom export functionality.  I have added the HighChart::$varName option which allows you to set the variable name without affecting the renderTo variable.